consolidate event reminder funcs
