+++ charm-tracker/imports/ui/pages/Calendar/TasksSidebar.js
@@ -1,5 +1,5 @@
 import React from "react"
-import GenerateTasksFromEvent from "../../../api/TaskHandler.js"
+import GenerateTasksFromEvent from "../../../api/taskHandling/TaskHandler.js"
 
 const TasksSidebar = props => {
   const eventWanted = "FZADqzACH2Qgoc2Ns"
